+++ Src/SOS.Observations.Api/Controllers/ObservationsController.cs
@@ -247,6 +247,7 @@ namespace SOS.Observations.Api.Controllers
         /// <param name="validateSearchFilter">If true, validation of search filter values will be made. I.e. HTTP bad request response will be sent if there are invalid parameter values.</param>
         /// <param name="translationCultureCode">Culture code used for vocabulary translation (sv-SE, en-GB).</param>
         /// <param name="protectedObservations">If true, only protected observations will be searched (this requires authentication and authorization). If false, public available observations will be searched.</param>
+        /// <param name="outputFormat">Select output format: JSON, GeoJSON with hierarchical properties, GeoJSON with flattened properties.</param>
         /// <returns>List of observations matching the provided search filter.</returns>
         /// <example>
         ///     Get all observations within 100m of provided point
@@ -283,7 +284,6 @@ namespace SOS.Observations.Api.Controllers
                     ValidatePropertyExists(nameof(sortBy), sortBy),
                     ValidateTranslationCultureCode(translationCultureCode));
                 if (validationResult.IsFailure) return BadRequest(validationResult.Error);
-
                 SearchFilter searchFilter = filter.ToSearchFilter(translationCultureCode, protectedObservations);
                 var result = await ObservationManager.GetChunkAsync(authorizationApplicationIdentifier, searchFilter, skip, take, sortBy, sortOrder);
                 PagedResultDto<dynamic> dto = result.ToPagedResultDto(result.Records);
@@ -636,6 +636,7 @@ namespace SOS.Observations.Api.Controllers
         /// <param name="validateSearchFilter">If true, validation of search filter values will be made. I.e. HTTP bad request response will be sent if there are invalid parameter values.</param>
         /// <param name="translationCultureCode">Culture code used for vocabulary translation (sv-SE, en-GB).</param>
         /// <param name="protectedObservations">If true, only protected observations will be searched (this requires authentication and authorization). If false, public available observations will be searched.</param>
+        /// <param name="outputFormat">Select output format: JSON, GeoJSON with hierarchical properties, GeoJSON with flattened properties.</param>
         /// <returns>List of observations matching the provided search filter.</returns>
         /// <example>
         ///     Get all observations within 100m of provided point
@@ -649,7 +650,7 @@ namespace SOS.Observations.Api.Controllers
         ///     }
         /// </example>
         [HttpPost("Internal/Search")]
-        [ProducesResponseType(typeof(PagedResultDto<Observation>), (int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(GeoPagedResultDto<Observation>), (int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
         [ProducesResponseType((int)HttpStatusCode.Unauthorized)]
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
@@ -663,7 +664,8 @@ namespace SOS.Observations.Api.Controllers
             [FromQuery] SearchSortOrder sortOrder = SearchSortOrder.Asc,
             [FromQuery] bool validateSearchFilter = false,
             [FromQuery] string translationCultureCode = "sv-SE",
-            [FromQuery] bool protectedObservations = false)
+            [FromQuery] bool protectedObservations = false,
+            [FromQuery] OutputFormatDto outputFormat = OutputFormatDto.Json)
         {
             try
             {
@@ -674,9 +676,12 @@ namespace SOS.Observations.Api.Controllers
                     ValidateTranslationCultureCode(translationCultureCode));
 
                 if (validationResult.IsFailure) return BadRequest(validationResult.Error);
-
+                if (outputFormat == OutputFormatDto.GeoJson || outputFormat == OutputFormatDto.GeoJsonFlat)
+                {
+                    filter.OutputFields = EnsureCoordinatesIsRetrievedFromDb(filter?.OutputFields);
+                }
                 var result = await ObservationManager.GetChunkAsync(authorizationApplicationIdentifier, filter.ToSearchFilterInternal(translationCultureCode, protectedObservations), skip, take, sortBy, sortOrder);
-                PagedResultDto<dynamic> dto = result.ToPagedResultDto(result.Records);
+                GeoPagedResultDto<dynamic> dto = result.ToGeoPagedResultDto(result.Records, outputFormat);
                 return new OkObjectResult(dto);
             }
             catch (AuthenticationRequiredException e)
@@ -688,6 +693,33 @@ namespace SOS.Observations.Api.Controllers
                 _logger.LogError(e, "SearchInternal error");
                 return new StatusCodeResult((int)HttpStatusCode.InternalServerError);
             }
+        }
+
+        private IEnumerable<string> EnsureCoordinatesIsRetrievedFromDb(IEnumerable<string> outputFields)
+        {
+            if (outputFields == null || !outputFields.Any())
+            {
+                return outputFields;
+            }
+
+            if (outputFields.Contains("location", StringComparer.OrdinalIgnoreCase) || (
+                outputFields.Contains("location.decimalLongitude", StringComparer.OrdinalIgnoreCase) &&
+                outputFields.Contains("location.decimalLatitude", StringComparer.OrdinalIgnoreCase)))
+            {
+                return outputFields;
+            }
+
+            var newList = new List<string>(outputFields);
+            if (!newList.Contains("location.decimalLongitude", StringComparer.OrdinalIgnoreCase))
+            {
+                newList.Add("location.decimalLongitude");
+            }
+            if (!newList.Contains("location.decimalLatitude", StringComparer.OrdinalIgnoreCase))
+            {
+                newList.Add("location.decimalLatitude");
+            }
+
+            return newList;
         }
 
         /// <summary>